+++ client/documents/mutations/index.ts
@@ -1,0 +1,1 @@
+export * from './updateProfile';
