+++ client/documents/queries/currentUser.ts
@@ -1,0 +1,21 @@
+import { graphql } from '@/graphql/gql';
+
+export const CurrentUserDocument = graphql(`
+  query CurrentUser {
+    currentUser {
+      id
+      uid
+      createdAt
+      updatedAt
+      profile {
+        id
+        name
+        birthDate
+        gender
+        height
+        weight
+        activityLevel
+      }
+    }
+  }
+`);
