+++ client/documents/queries/index.ts
@@ -1,0 +1,2 @@
+export * from './currentUser';
+export * from './profileEditCurrentUser';
